added calculator endpoint for clients to check length of query, fixed edit cart bug with discount
