+++ apps/api/internal/input/calculator.go
@@ -1,0 +1,7 @@
+package input
+
+type CallbackCalculatorQueryInput struct {
+	Category    string `query:"category"`
+	Subcategory string `query:"subcategory"`
+	ProductName string `query:"productName"`
+}
